requirements.txt
requests
beautifulsoup4
google-generativeai
edge-tts
asyncio